Bug 1932969 - Use doubles for intermediate values to prevent float overflow r=dholbert
Differential Revision: https://phabricator.services.mozilla.com/D230025

+++ dom/svg/SVGViewBoxSMILType.cpp
@@ -90,10 +90,10 @@ nsresult SVGViewBoxSMILType::ComputeDistance(const SMILValue& aFrom,
   // given change in the left side as it is for an equal change in the opposite
   // side. See https://bugzilla.mozilla.org/show_bug.cgi?id=541884#c12
 
-  float dLeft = to->x - from->x;
-  float dTop = to->y - from->y;
-  float dRight = (to->x + to->width) - (from->x + from->width);
-  float dBottom = (to->y + to->height) - (from->y + from->height);
+  double dLeft = to->x - from->x;
+  double dTop = to->y - from->y;
+  double dRight = (to->x + to->width) - (from->x + from->width);
+  double dBottom = (to->y + to->height) - (from->y + from->height);
 
   aDistance = std::sqrt(dLeft * dLeft + dTop * dTop + dRight * dRight +
                         dBottom * dBottom);
